+++ test/libsolidity/syntaxTests/pragma/unterminated_pragma.sol
@@ -1,0 +1,3 @@
+pragma solidity 0.4.3
+// ----
+// ParserError 2314: (22-22): Expected ';' but got end of source